api/serializers.py
from rest_framework import serializers
from .models import models
from blog.models import Post


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = '__all__'
